Changed test

+++ LR4/calc.cpp
@@ -8,12 +8,12 @@ int main() {
 	std::cin >> num_2;
 	std::cout << "Type the third number: ";
 	std::cin >> num_3;
-	if ((num_1 >= 0) and (num_2 >= 0) and (num_3 >=0) and (num_1 > num_2)) {
+	if ((num_1 >= 0) and (num_2 >= 0) and (num_3 >= 0) and (num_1 > num_2)) {
 		std::cout << num_1 << " + " << num_2 << " = " << num_1 + num_2 << std::endl;
 		std::cout << num_1 << " - " << num_2 << " = " << num_1 - num_2 << std::endl;
 		std::cout << num_1 << " * " << num_2 << " = " << num_1 * num_2 << std::endl;
 		std::cout << num_1 << " / " << num_2 << " = " << num_1 / num_2 << std::endl;
-		std::cout << num_1 << ".0 / " << num_2 << ".0 = " << double(num_1) / double(num_2) << std::endl;
+		std::cout << num_1 << ".0 / " << num_2 << ".0 = " << float(num_1) / float(num_2) << std::endl;
 		std::cout << num_1 << " % " << num_2 << " = " << num_1 % num_2 << std::endl;
 		std::cout << num_1 << " + " << num_2 << " * " << num_3 << " = " << num_1 + num_2 * num_3 << std::endl;
 		std::cout << "(" << num_1 << " + " << num_2 << ") * " << num_3 << " = " << (num_1 + num_2) * num_3 << std::endl;
